Made poster in details view static and fixed info view
Poster in details was part of scroll view due to which
it was scrolling too. Keeping it static to match the
provided example gif's outlook.

Info view which is now part of scroll view only has
title and overview of movie and it will bounce off on
scrolling just like its being done in the gif given
in assignment's problem statement.

Signed-off-by: Usman Ajmal <syedusmanajmal@gmail.com>