Show low resolution image first for poster and then high
On loading details view for a poster, first its low
resolution image is loaded which appears very fast as
size of image is slow. Then, in its success callback
we load higher resoution image which takes time to load.

This provides user better experience as it increased
"perceived speed" of the app.

Signed-off-by: Usman Ajmal <syedusmanajmal@gmail.com>